requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
altgraph==0.17.5
    # via pyinstaller
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6
    # via tqdm
gemmi==0.7.4
    # via -r requirements.in
idna==3.11
    # via requests
numpy==2.4.2
    # via
    #   -r requirements.in
    #   pandas
    #   rdkit
packaging==26.0
    # via
    #   pyinstaller
    #   pyinstaller-hooks-contrib
pandas==3.0.0
    # via -r requirements.in
pefile==2024.8.26
    # via pyinstaller
pillow==12.1.0
    # via rdkit
pyinstaller==6.18.0
    # via -r requirements.in
pyinstaller-hooks-contrib==2026.0
    # via pyinstaller
pyqt6==6.10.2
    # via -r requirements.in
pyqt6-qt6==6.10.2
    # via pyqt6
pyqt6-sip==13.11.0
    # via pyqt6
python-dateutil==2.9.0.post0
    # via pandas
pywin32-ctypes==0.2.3
    # via pyinstaller
pyyaml==6.0.3
    # via -r requirements.in
rdkit==2025.9.4
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
tqdm==4.67.3
    # via -r requirements.in
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
